hotfix missing original entity
Since new entities are send from the `PrePersist` event, the
unit of work no longer marks them as scheduled for insert. Checking
the original for `null` will provide you with the same information
in a much more elegant way and fixes the problem of having no
original when using the `STRATEGY_COPY_PREVIOUS`.